mongodb

+++ server/src/controllers/todos.controller.ts
@@ -3,14 +3,15 @@ import {
   Controller,
   Delete,
   Get,
+  HttpException,
   Param,
-  ParseIntPipe,
   Post,
   Put,
 } from '@nestjs/common';
-import { Todo } from '../interfaces/todo.interface';
 import { CreateTodoDto } from '../dto/create-todo.dto';
 import { TodosService } from '../services/todos.service';
+import { Todo } from '../schemas/todo.schema';
+import mongoose from 'mongoose';
 import { UpdateTodoDto } from '../dto/update-todo.dto';
 
 @Controller('todos')
@@ -23,8 +24,12 @@ export class TodosController {
   }
 
   @Get(':id')
-  async getTodoById(@Param('id', ParseIntPipe) id: number): Promise<Todo> {
-    return this.todosService.getTodoById(id);
+  async getTodoById(@Param('id') id: string) {
+    const isValid = mongoose.Types.ObjectId.isValid(id);
+    if (!isValid) throw new HttpException('Invalid ID', 400);
+    const foundTodo = this.todosService.getTodoById(id);
+    if (!foundTodo) throw new HttpException('Todo not found', 404);
+    return foundTodo;
   }
 
   @Post()
@@ -34,14 +39,18 @@ export class TodosController {
 
   @Put(':id')
   async updateTodo(
-    @Param('id', ParseIntPipe) id: number,
+    @Param('id') id: string,
     @Body() updateTodoDto: UpdateTodoDto,
   ): Promise<Todo> {
+    const isValid = mongoose.Types.ObjectId.isValid(id);
+    if (!isValid) throw new HttpException('Invalid ID', 400);
     return this.todosService.updateTodo(id, updateTodoDto);
   }
 
   @Delete(':id')
-  async deleteTodo(@Param('id', ParseIntPipe) id: number): Promise<Todo> {
+  async deleteTodo(@Param('id') id: string): Promise<Todo> {
+    const isValid = mongoose.Types.ObjectId.isValid(id);
+    if (!isValid) throw new HttpException('Invalid ID', 400);
     return this.todosService.deleteTodo(id);
   }
 }
